Spatial Clustering of GInteractions

Merges spatially proximal chromatin loops into consensus interactions using graph-based clustering. Loops are considered overlapping if both anchors are within gap bp of each other. Each resulting cluster is represented by the union genomic range (min start to max end) spanning all its members.

Usage

reduce_ginteractions(gi, gap = 1000)

Arguments

gi

A GInteractions object.

gap

Numeric. Maximum distance (in base pairs) allowed between anchors to consider two loops overlapping. Default: 1000.

Value

A list with two elements:

gi

Reduced GInteractions object, one per cluster.

membership

Integer vector indicating cluster assignment for each input loop.

Metadata columns include cluster_id, n_members, and averaged score.

Examples

# 1. Load example data (loops that are close to each other)
bedpe_path <- system.file("extdata", "example_loops_1.bedpe", package = "looplook")

if (bedpe_path != "") {
  # Convert BEDPE to GInteractions object
  gi_raw <- bedpe_to_gi(bedpe_path)

  # 2. Run clustering
  # Merge loops if their anchors are within 1000bp
  res <- reduce_ginteractions(gi_raw, gap = 1000)

  # 3. Inspect results
  # The 'gi' element contains the merged consensus loops
  print(res$gi)

  # The 'membership' vector tells which original loop belongs to which cluster
  head(res$membership)

  # Check cluster sizes (how many loops were merged into each cluster)
  table(res$membership)
}
#> StrictGInteractions object with 1000 interactions and 4 metadata columns:
#>          seqnames1             ranges1     seqnames2             ranges2 |
#>              <Rle>           <IRanges>         <Rle>           <IRanges> |
#>      [1]      chr1 112721687-112724586 ---      chr1 112791557-112793600 |
#>      [2]      chr1 113867048-113875597 ---      chr1 113927096-113933645 |
#>      [3]      chr1 115060199-115071191 ---      chr1 116162680-116169369 |
#>      [4]      chr1 118807309-118811756 ---      chr1 119277565-119282472 |
#>      [5]      chr1 112399375-112406783 ---      chr1 113756152-113762298 |
#>      ...       ...                 ... ...       ...                 ... .
#>    [996]      chr1   12374800-12386079 ---      chr1   12848181-12853963 |
#>    [997]      chr1 109654027-109657363 ---      chr1 110405568-110409319 |
#>    [998]      chr1 115176997-115186579 ---      chr1 115779639-115785509 |
#>    [999]      chr1   12374800-12386079 ---      chr1   12548119-12551309 |
#>   [1000]      chr1   10054549-10058953 ---      chr1   11801550-11808331 |
#>          cluster_id n_members     score    n_reps
#>           <numeric> <integer> <numeric> <integer>
#>      [1]          1         1         1         1
#>      [2]          2         1         4         1
#>      [3]          3         1         2         1
#>      [4]          4         1         1         1
#>      [5]          5         1         2         1
#>      ...        ...       ...       ...       ...
#>    [996]        996         1         1         1
#>    [997]        997         1         1         1
#>    [998]        998         1         1         1
#>    [999]        999         1         2         1
#>   [1000]       1000         1         1         1
#>   -------
#>   regions: 689 ranges and 0 metadata columns
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#>    1    2    3    4    5    6    7    8    9   10   11   12   13   14   15   16 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>   17   18   19   20   21   22   23   24   25   26   27   28   29   30   31   32 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>   33   34   35   36   37   38   39   40   41   42   43   44   45   46   47   48 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>   49   50   51   52   53   54   55   56   57   58   59   60   61   62   63   64 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>   65   66   67   68   69   70   71   72   73   74   75   76   77   78   79   80 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>   81   82   83   84   85   86   87   88   89   90   91   92   93   94   95   96 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>   97   98   99  100  101  102  103  104  105  106  107  108  109  110  111  112 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  113  114  115  116  117  118  119  120  121  122  123  124  125  126  127  128 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  129  130  131  132  133  134  135  136  137  138  139  140  141  142  143  144 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  145  146  147  148  149  150  151  152  153  154  155  156  157  158  159  160 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  161  162  163  164  165  166  167  168  169  170  171  172  173  174  175  176 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  177  178  179  180  181  182  183  184  185  186  187  188  189  190  191  192 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  193  194  195  196  197  198  199  200  201  202  203  204  205  206  207  208 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  209  210  211  212  213  214  215  216  217  218  219  220  221  222  223  224 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  225  226  227  228  229  230  231  232  233  234  235  236  237  238  239  240 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  241  242  243  244  245  246  247  248  249  250  251  252  253  254  255  256 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  257  258  259  260  261  262  263  264  265  266  267  268  269  270  271  272 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  273  274  275  276  277  278  279  280  281  282  283  284  285  286  287  288 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  289  290  291  292  293  294  295  296  297  298  299  300  301  302  303  304 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  305  306  307  308  309  310  311  312  313  314  315  316  317  318  319  320 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  321  322  323  324  325  326  327  328  329  330  331  332  333  334  335  336 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  337  338  339  340  341  342  343  344  345  346  347  348  349  350  351  352 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  353  354  355  356  357  358  359  360  361  362  363  364  365  366  367  368 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  369  370  371  372  373  374  375  376  377  378  379  380  381  382  383  384 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  385  386  387  388  389  390  391  392  393  394  395  396  397  398  399  400 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  401  402  403  404  405  406  407  408  409  410  411  412  413  414  415  416 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  417  418  419  420  421  422  423  424  425  426  427  428  429  430  431  432 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  433  434  435  436  437  438  439  440  441  442  443  444  445  446  447  448 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  449  450  451  452  453  454  455  456  457  458  459  460  461  462  463  464 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  465  466  467  468  469  470  471  472  473  474  475  476  477  478  479  480 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  481  482  483  484  485  486  487  488  489  490  491  492  493  494  495  496 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  497  498  499  500  501  502  503  504  505  506  507  508  509  510  511  512 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  513  514  515  516  517  518  519  520  521  522  523  524  525  526  527  528 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  529  530  531  532  533  534  535  536  537  538  539  540  541  542  543  544 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  545  546  547  548  549  550  551  552  553  554  555  556  557  558  559  560 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  561  562  563  564  565  566  567  568  569  570  571  572  573  574  575  576 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  577  578  579  580  581  582  583  584  585  586  587  588  589  590  591  592 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  593  594  595  596  597  598  599  600  601  602  603  604  605  606  607  608 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  609  610  611  612  613  614  615  616  617  618  619  620  621  622  623  624 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  625  626  627  628  629  630  631  632  633  634  635  636  637  638  639  640 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  641  642  643  644  645  646  647  648  649  650  651  652  653  654  655  656 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  657  658  659  660  661  662  663  664  665  666  667  668  669  670  671  672 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  673  674  675  676  677  678  679  680  681  682  683  684  685  686  687  688 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  689  690  691  692  693  694  695  696  697  698  699  700  701  702  703  704 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  705  706  707  708  709  710  711  712  713  714  715  716  717  718  719  720 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  721  722  723  724  725  726  727  728  729  730  731  732  733  734  735  736 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  737  738  739  740  741  742  743  744  745  746  747  748  749  750  751  752 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  753  754  755  756  757  758  759  760  761  762  763  764  765  766  767  768 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  769  770  771  772  773  774  775  776  777  778  779  780  781  782  783  784 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  785  786  787  788  789  790  791  792  793  794  795  796  797  798  799  800 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  801  802  803  804  805  806  807  808  809  810  811  812  813  814  815  816 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  817  818  819  820  821  822  823  824  825  826  827  828  829  830  831  832 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  833  834  835  836  837  838  839  840  841  842  843  844  845  846  847  848 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  849  850  851  852  853  854  855  856  857  858  859  860  861  862  863  864 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  865  866  867  868  869  870  871  872  873  874  875  876  877  878  879  880 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  881  882  883  884  885  886  887  888  889  890  891  892  893  894  895  896 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  897  898  899  900  901  902  903  904  905  906  907  908  909  910  911  912 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  913  914  915  916  917  918  919  920  921  922  923  924  925  926  927  928 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  929  930  931  932  933  934  935  936  937  938  939  940  941  942  943  944 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  945  946  947  948  949  950  951  952  953  954  955  956  957  958  959  960 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  961  962  963  964  965  966  967  968  969  970  971  972  973  974  975  976 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  977  978  979  980  981  982  983  984  985  986  987  988  989  990  991  992 
#>    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
#>  993  994  995  996  997  998  999 1000 
#>    1    1    1    1    1    1    1    1